refactor: clean comments

+++ src/modules/backlight.h
@@ -5,7 +5,8 @@
 
 extern BrightnessMode currentBrightnessMode;
 
-constexpr uint32_t ALS_MEASUREMENT_INTERVAL_MS = 8000;  // ALS 測定間隔 [ms]
+// ALS 測定間隔 [ms]
+constexpr uint16_t ALS_MEASUREMENT_INTERVAL_MS = 8000;
 
 void updateBacklightLevel();
